feat: LocalPath also sets group (if not specified) based on user (#38)
This PR aligns the behaviour of the file creation methods (mkdir, write_bytes,
write_text) when the user argument is provided but the group argument is not.

In this case, Pebble sets the group based on the user (like chown $user:
$path), while shutil.chown by default does not (like chown $user $path). Since
changing the Pebble behaviour to match shutil here would require additional API
calls (and possibly creating the directory twice?), and since the user and
group arguments for the file creation methods are an API design that follows
Pebble rather than pathlib, LocalPath is updated to follow Pebble's behaviour
here.

This PR updates the unit tests to reflect this change, and adds a collection of
integration tests for file ownership.

Documentation of the error case when group is provided without user for
ContainerPath has also been added.

Resolves #39

+++ pathops/tests/integration/juju/charms/machine/src/charm.py
@@ -15,8 +15,12 @@
 
 """Charm the application."""
 
+from __future__ import annotations
+
 import logging
 import shutil
+import subprocess
+import typing
 
 import common
 import ops
@@ -25,6 +29,9 @@ import ops
 #       after next pyright release fixes:
 #       https://github.com/microsoft/pyright/issues/10203
 import charmlibs.pathops as pathops
+
+if typing.TYPE_CHECKING:
+    from typing import Sequence
 
 logger = logging.getLogger(__name__)
 
@@ -43,11 +50,17 @@ class Charm(common.Charm):
 
     def remove_path(self, path: pathops.PathProtocol, recursive: bool = False) -> None:
         assert isinstance(path, pathops.LocalPath)
+        if not path.exists():
+            return
         if not path.is_dir():
             return path.unlink()
         if not recursive:
             return path.rmdir()
         shutil.rmtree(path)
+
+    def exec(self, cmd: Sequence[str]) -> int:
+        process = subprocess.run(cmd)
+        return process.returncode
 
 
 if __name__ == '__main__':  # pragma: nocover
